+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# ggpattern 0.1.2  2020-06-28
+
+* Fix issue with attempts to fill an area with a pattern when it has zero area. (Issue #9)
+
 # ggpattern 0.1.1  2020-04-23
 
 * Use `magick::image_data()` instead of `magick::as_EBImage()` for internal